+++ receiver/swohostmetricsreceiver/internal/providers/firewall/provider_windows.go
@@ -20,7 +20,7 @@ import (
 	"sync"
 
 	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/providers"
-	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/registry"
+	"github.com/solarwinds/solarwinds-otel-collector-contrib/tools/registry"
 	"go.uber.org/zap"
 )