+++ receiver/swohostmetricsreceiver/internal/providers/installedsoftware/provider_windows.go
@@ -17,7 +17,7 @@ package installedsoftware
 import (
 	"fmt"
 
-	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/registry"
+	"github.com/solarwinds/solarwinds-otel-collector-contrib/tools/registry"
 	"go.uber.org/zap"
 )