+++ receiver/swohostmetricsreceiver/internal/providers/loggedusers/provider_windows.go
@@ -16,7 +16,7 @@ package loggedusers
 
 import (
 	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/providers"
-	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/registry"
+	"github.com/solarwinds/solarwinds-otel-collector-contrib/tools/registry"
 )
 
 type provider struct {