+++ receiver/swohostmetricsreceiver/internal/providers/timezone/provider_windows.go
@@ -16,7 +16,7 @@ package timezone
 
 import (
 	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/providers"
-	"github.com/solarwinds/solarwinds-otel-collector-contrib/receiver/swohostmetricsreceiver/internal/wmi"
+	"github.com/solarwinds/solarwinds-otel-collector-contrib/tools/wmi"
 )
 
 type provider struct {